# [PERF] stock: only force prefetch for non-cached records