--- a/addons/mrp/tests/test_order.py
+++ b/addons/mrp/tests/test_order.py
@@ -1549,6 +1549,7 @@ class TestMrpOrder(TestMrpCommon):
         (move1 | move2 | move3)._action_confirm()
 
         mo.invalidate_cache(['components_availability', 'components_availability_state'], mo.ids)
+        mo.move_raw_ids.invalidate_cache(['forecast_availability', 'forecast_expected_date'], mo.move_raw_ids.ids)
         self.assertEqual(mo.components_availability, f'Exp {format_date(self.env, tommorrow)}')
         self.assertEqual(mo.components_availability_state, 'late')
 
@@ -1561,6 +1562,7 @@ class TestMrpOrder(TestMrpCommon):
         (move1 | move2 | move3)._action_done()
 
         mo.invalidate_cache(['components_availability', 'components_availability_state'], mo.ids)
+        mo.move_raw_ids.invalidate_cache(['forecast_availability', 'forecast_expected_date'], mo.move_raw_ids.ids)
         self.assertEqual(mo.components_availability, 'Available')
         self.assertEqual(mo.components_availability_state, 'available')