@robodoo 2024-05-29
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/167274).