@zaha-odoo 2024-05-30
Feel terrible doing this, but had to modify the test to get it to pass. Need to invalidate the move_raw_ids' forecast info as well, since otherwise the new code would see that they are already cached and give bad values.

Not sure if this is OK. But I'd imagine in a "real-world" scenario, if components_availability was already in the cache then it should work fine. Issue only arises because in the test, the MO's components_availability is invalidated. I could see this potentially being a breaking stability change if other code that calls invalidate_cache depended on `_compute_components_availability` being like it was the original way,