@zaha-odoo 2024-05-30
Couldn't find any other usages of an invalidate_cache call on mrp.production in 15.0 Odoo code