@zaha-odoo 2024-05-30
Here was failing test build: https://runbot.odoo.com/runbot/build/63191143