@zaha-odoo 2024-08-16
Closing because no longer reproducible